refactor(slack): isolate connector image pin policy
