+++ apps/slack/cmd/main.go
@@ -7,7 +7,6 @@ package main
 
 import (
 	"context"
-	_ "crypto/sha256" // register sha256 for github.com/distribution/reference digest parsing
 	"errors"
 	"fmt"
 	"log/slog"
@@ -22,10 +21,9 @@ import (
 	"syscall"
 	"time"
 
-	dockerref "github.com/distribution/reference"
-
 	"github.com/layervai/qurl-integrations/apps/slack/internal"
 	"github.com/layervai/qurl-integrations/apps/slack/internal/agent"
+	"github.com/layervai/qurl-integrations/apps/slack/internal/connectorimage"
 	"github.com/layervai/qurl-integrations/apps/slack/internal/oauth"
 	"github.com/layervai/qurl-integrations/apps/slack/internal/slackdata"
 	"github.com/layervai/qurl-integrations/apps/slack/internal/slackinstall"
@@ -1042,41 +1040,41 @@ func readTunnelImageConfig() (string, error) {
 	if image != "" {
 		// Keep startup errors explicit: they land in operator logs, and each
 		// branch carries the remediation so bad image config cannot be masked.
-		switch classifyTunnelImagePin(image) {
-		case tunnelImagePinned:
+		switch connectorimage.ClassifyPin(image) {
+		case connectorimage.Pinned:
 			return image, nil
-		case tunnelImageLatestDigest:
+		case connectorimage.LatestDigest:
 			return "", fmt.Errorf(
 				"%s digest pins must not include a latest tag; use an image@sha256:<64 lowercase hex> digest without :latest, or a specific non-latest release tag before the digest; %s",
 				envQURLConnectorImage, connectorImageFallbackHint,
 			)
-		case tunnelImageUppercaseDigest:
+		case connectorimage.UppercaseDigest:
 			return "", fmt.Errorf(
 				"%s digest must use 64 lowercase hex characters after sha256:; %s",
 				envQURLConnectorImage, connectorImageFallbackHint,
 			)
-		case tunnelImageMalformedReference:
+		case connectorimage.MalformedReference:
 			return "", fmt.Errorf(
 				"%s image references must use image@sha256:<64 lowercase hex> with a full lowercase image name, or a full lowercase image name with a single non-empty, non-latest tag; %s",
 				envQURLConnectorImage, connectorImageFallbackHint,
 			)
-		case tunnelImageAmbiguousReference:
+		case connectorimage.AmbiguousReference:
 			return "", fmt.Errorf(
 				"%s slashless registry references must include a repository path, for example gcr.io/<org>/<image>:v1, or use image@sha256:<64 lowercase hex> digest; %s",
 				envQURLConnectorImage, connectorImageFallbackHint,
 			)
-		case tunnelImageMalformedDigest:
+		case connectorimage.MalformedDigest:
 			return "", fmt.Errorf(
 				"%s digest references must use image@sha256:<64 lowercase hex> with a full image name; %s",
 				envQURLConnectorImage, connectorImageFallbackHint,
 			)
-		case tunnelImageFloating:
+		case connectorimage.Floating:
 			return "", fmt.Errorf(
 				"%s must be pinned: use a non-latest tag or image@sha256:<64 lowercase hex> digest; %s",
 				envQURLConnectorImage, connectorImageFallbackHint,
 			)
 		}
-		// Future tunnelImagePinStatus values must fail closed.
+		// Future connectorimage.PinStatus values must fail closed.
 		return "", fmt.Errorf("%s could not validate image pinning; %s", envQURLConnectorImage, connectorImageFallbackHint)
 	}
 
@@ -1090,237 +1088,6 @@ func readTunnelImageConfig() (string, error) {
 	default:
 		return "", fmt.Errorf("%s=%q is unsupported; set %s only for dev/sandbox, or set %s to a specific non-latest tag or digest", envQURLConnectorImageFallback, rawFallback, connectorImageFallbackOptIn, envQURLConnectorImage)
 	}
-}
-
-type tunnelImagePinStatus int
-
-const (
-	tunnelImagePinned tunnelImagePinStatus = iota
-	tunnelImageFloating
-	tunnelImageLatestDigest
-	tunnelImageUppercaseDigest
-	tunnelImageMalformedReference
-	tunnelImageAmbiguousReference
-	tunnelImageMalformedDigest
-)
-
-// classifyTunnelImagePin decides whether an operator-provided image ref is
-// pinned enough for production startup and, if not, which startup error should
-// explain it. Docker reference grammar comes from github.com/distribution/reference;
-// the local guards are qURL policy: reject floating/latest forms, reject
-// slashless registry-looking names that Docker would normalize to Hub library
-// images, and keep digest/tag remediation messages specific.
-func classifyTunnelImagePin(image string) tunnelImagePinStatus {
-	name, digest, hasDigest := strings.Cut(image, "@")
-	if hasDigest {
-		return classifyDigestImagePin(name, digest)
-	}
-	return classifyTaggedImagePin(name)
-}
-
-func classifyDigestImagePin(name, digest string) tunnelImagePinStatus {
-	lastSlash := strings.LastIndex(name, "/")
-	lastColon := strings.LastIndex(name, ":")
-	nameSuffix := name[lastSlash+1:]
-	repositoryName := imageRepositoryName(name, lastSlash, lastColon)
-
-	if name == "" {
-		return tunnelImageMalformedDigest
-	}
-	if lastColon == lastSlash+1 {
-		return tunnelImageMalformedReference
-	}
-	if imageNameHasUppercase(repositoryName) {
-		return tunnelImageMalformedReference
-	}
-	// The digest controls image resolution, but reject
-	// repo:latest@sha256:<hex> anyway so customer-facing snippets never
-	// visibly advertise the floating latest tag.
-	if imageNameHasLatestTag(name) {
-		return tunnelImageLatestDigest
-	}
-	if strings.HasSuffix(nameSuffix, ":") || strings.Count(nameSuffix, ":") > 1 {
-		return tunnelImageMalformedReference
-	}
-	digestStatus := classifySHA256ImageDigest(digest)
-	switch digestStatus {
-	case sha256DigestUppercaseHex:
-		return tunnelImageUppercaseDigest
-	case sha256DigestInvalid:
-		return tunnelImageMalformedDigest
-	case sha256DigestValid:
-		// Continue into the Docker reference parser below.
-	default:
-		// Future sha256DigestStatus values must fail closed.
-		return tunnelImageMalformedDigest
-	}
-	if strings.EqualFold(repositoryName, "sha256") || slashlessRegistryReference(name, lastSlash, lastColon) {
-		return tunnelImageMalformedDigest
-	}
-	if !isParsedDockerReferencePinned(name + "@" + digest) {
-		return tunnelImageMalformedDigest
-	}
-	return tunnelImagePinned
-}
-
-func classifyTaggedImagePin(name string) tunnelImagePinStatus {
-	lastSlash := strings.LastIndex(name, "/")
-	lastColon := strings.LastIndex(name, ":")
-
-	if lastColon <= lastSlash {
-		return tunnelImageFloating
-	}
-	// Non-digest refs surface malformed tag syntax before latest-specific
-	// guidance. Digest refs check latest first so repo:latest@sha256:... never
-	// renders a customer snippet that visibly advertises :latest.
-	nameSuffix := name[lastSlash+1:]
-	if strings.Count(nameSuffix, ":") > 1 {
-		return tunnelImageMalformedReference
-	}
-	if lastSlash < 0 && looksLikeRegistryHost(name[:lastColon]) {
-		return tunnelImageAmbiguousReference
-	}
-	tag := name[lastColon+1:]
-	if tag == "" {
-		return tunnelImageMalformedReference
-	}
-	if imageNameHasLatestTag(name) {
-		return tunnelImageFloating
-	}
-	repositoryName := imageRepositoryName(name, lastSlash, lastColon)
-	// Avoid treating bare digest-looking sha256:<hex> input as an image named
-	// "sha256" with a tag.
-	if strings.EqualFold(repositoryName, "sha256") {
-		return tunnelImageMalformedDigest
-	}
-	if lastColon == lastSlash+1 || imageNameHasUppercase(repositoryName) || !isParsedDockerReferencePinned(name) {
-		return tunnelImageMalformedReference
-	}
-	// Non-latest tags are trusted release labels by operator convention; use
-	// image@sha256:<digest> when byte-for-byte image immutability is required.
-	return tunnelImagePinned
-}
-
-type sha256DigestStatus int
-
-const (
-	sha256DigestValid sha256DigestStatus = iota
-	sha256DigestUppercaseHex
-	sha256DigestInvalid
-)
-
-func classifySHA256ImageDigest(digest string) sha256DigestStatus {
-	const prefix = "sha256:"
-	if !strings.HasPrefix(digest, prefix) {
-		return sha256DigestInvalid
-	}
-	hex := strings.TrimPrefix(digest, prefix)
-	if len(hex) != 64 {
-		return sha256DigestInvalid
-	}
-	hasUpper := false
-	for _, r := range hex {
-		switch {
-		case r >= '0' && r <= '9':
-		case r >= 'a' && r <= 'f':
-		case r >= 'A' && r <= 'F':
-			hasUpper = true
-		default:
-			return sha256DigestInvalid
-		}
-	}
-	// Image digests produced by GHCR/Docker for this connector are sha256.
-	// Reject other OCI digest algorithms until we have a concrete need for
-	// them, rather than silently widening the accepted operator input.
-	// Require canonical lowercase rather than normalizing so operator config
-	// stays byte-for-byte identical to the pinned digest shown to customers.
-	if hasUpper {
-		return sha256DigestUppercaseHex
-	}
-	return sha256DigestValid
-}
-
-func imageNameHasLatestTag(name string) bool {
-	components := strings.Split(name, "/")
-	for i, component := range components {
-		if i == 0 && len(components) > 1 && !firstPathComponentHasValidPort(component) {
-			continue
-		}
-		for _, tag := range strings.Split(component, ":")[1:] {
-			if strings.EqualFold(tag, "latest") {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func imageRepositoryName(name string, lastSlash, lastColon int) string {
-	if lastColon > lastSlash {
-		taglessName := name[:lastColon]
-		if lastSlash < 0 && looksLikeRegistryHost(taglessName) {
-			return name
-		}
-		return taglessName
-	}
-	return name
-}
-
-func slashlessRegistryReference(name string, lastSlash, lastColon int) bool {
-	if lastSlash >= 0 {
-		return false
-	}
-	if lastColon > lastSlash {
-		return looksLikeRegistryHost(name[:lastColon])
-	}
-	return looksLikeRegistryHost(name)
-}
-
-func isParsedDockerReferencePinned(image string) bool {
-	parsed, err := dockerref.ParseAnyReference(image)
-	if err != nil {
-		return false
-	}
-	if _, ok := parsed.(dockerref.Named); !ok {
-		return false
-	}
-	if _, ok := parsed.(dockerref.Tagged); ok {
-		return true
-	}
-	_, ok := parsed.(dockerref.Digested)
-	return ok
-}
-
-func imageNameHasUppercase(name string) bool {
-	for _, r := range name {
-		if r >= 'A' && r <= 'Z' {
-			return true
-		}
-	}
-	return false
-}
-
-func firstPathComponentHasValidPort(component string) bool {
-	lastColon := strings.LastIndex(component, ":")
-	if lastColon < 0 {
-		return true
-	}
-	port := component[lastColon+1:]
-	if port == "" {
-		return false
-	}
-	for _, r := range port {
-		if r < '0' || r > '9' {
-			return false
-		}
-	}
-	return true
-}
-
-func looksLikeRegistryHost(host string) bool {
-	// Reserve localhost, dotted names, and host:port forms for registry hosts,
-	// while leaving bare words as valid Docker-style single-segment image names.
-	return strings.EqualFold(host, "localhost") || strings.ContainsAny(host, ".:")
 }
 
 // readPoolSizeEnv parses a pool-size env var. Empty is "use default" silently;
