+++ apps/slack/cmd/main_test.go
@@ -262,107 +262,6 @@ func TestRunValidatesTunnelImageBeforeInfraSetup(t *testing.T) {
 	}
 }
 
-func TestClassifyTunnelImagePin(t *testing.T) {
-	t.Parallel()
-	validDigest := "sha256:" + strings.Repeat("a", 64)
-	cases := []struct {
-		name  string
-		image string
-		want  tunnelImagePinStatus
-	}{
-		{name: "version tag", image: testConnectorVersionImage, want: tunnelImagePinned},
-		{name: "registry host with port and tag", image: "registry.example.com:5000/layervai/qurl-connector:v1.2.3", want: tunnelImagePinned},
-		{name: "uppercase non-latest tag", image: testConnectorImageRepo + ":V1", want: tunnelImagePinned},
-		{name: "uppercase non-latest tag with digest", image: testConnectorImageRepo + ":V1@" + validDigest, want: tunnelImagePinned},
-		{name: "digest", image: testConnectorImageRepo + "@" + validDigest, want: tunnelImagePinned},
-		{name: "registry host with port and digest", image: "registry.example.com:5000/layervai/qurl-connector@" + validDigest, want: tunnelImagePinned},
-		{name: "single segment image digest", image: "qurl-connector@" + validDigest, want: tunnelImagePinned},
-		{name: "version tag with digest", image: testConnectorVersionImage + "@" + validDigest, want: tunnelImagePinned},
-		{name: "latest tag with digest", image: testConnectorLatestImage + "@" + validDigest, want: tunnelImageLatestDigest},
-		{name: "uppercase latest tag with digest", image: testConnectorImageRepo + ":LATEST@" + validDigest, want: tunnelImageLatestDigest},
-		{name: "multi-colon latest tag with digest", image: testConnectorImageRepo + ":latest:v1@" + validDigest, want: tunnelImageLatestDigest},
-		{name: "path component latest tag with digest", image: "ghcr.io/foo:latest/qurl-connector@" + validDigest, want: tunnelImageLatestDigest},
-		{name: "blank image", want: tunnelImageFloating},
-		{name: "implicit latest", image: testConnectorImageRepo, want: tunnelImageFloating},
-		{name: "explicit latest", image: testConnectorLatestImage, want: tunnelImageFloating},
-		{name: "uppercase latest", image: testConnectorImageRepo + ":LATEST", want: tunnelImageFloating},
-		{name: "path component latest tag", image: "ghcr.io/foo:latest/qurl-connector:v1", want: tunnelImageFloating},
-		{name: "path component uppercase latest tag", image: "ghcr.io/foo:LATEST/qurl-connector:v1", want: tunnelImageFloating},
-		{name: "invalid latest-like registry port fails safe", image: "localhost:latest/qurl-connector:v1", want: tunnelImageMalformedReference},
-		{name: "multi-colon latest tag", image: testConnectorImageRepo + ":latest:v1", want: tunnelImageMalformedReference},
-		{name: "multi-colon non-latest tag", image: testConnectorImageRepo + ":v1:v2", want: tunnelImageMalformedReference},
-		{name: "slashless multi-colon non-latest tag", image: "repo:v1:v2", want: tunnelImageMalformedReference},
-		{name: "empty tag", image: testConnectorImageRepo + ":", want: tunnelImageMalformedReference},
-		{name: "non-numeric registry port", image: "host:abc/qurl-connector:v1", want: tunnelImageMalformedReference},
-		{name: "empty path component with tag", image: "ghcr.io//qurl-connector:v1", want: tunnelImageMalformedReference},
-		{name: "uppercase repository namespace", image: "ghcr.io/LayerV/qurl-connector:v1", want: tunnelImageMalformedReference},
-		{name: "uppercase registry host", image: "GHCR.IO/layervai/qurl-connector:v1", want: tunnelImageMalformedReference},
-		{name: "uppercase registry host with digest", image: "GHCR.IO/layervai/qurl-connector@" + validDigest, want: tunnelImageMalformedReference},
-		{name: "empty name with tag", image: ":v1", want: tunnelImageMalformedReference},
-		{name: "empty name with latest digest", image: ":latest@" + validDigest, want: tunnelImageMalformedReference},
-		{name: "registry port without tag", image: "localhost:5000/layervai/qurl-connector", want: tunnelImageFloating},
-		{name: "bare localhost port", image: "localhost:5000", want: tunnelImageAmbiguousReference},
-		{name: "mixed-case localhost port", image: "Localhost:5000", want: tunnelImageAmbiguousReference},
-		{name: "uppercase localhost port", image: "LOCALHOST:5000", want: tunnelImageAmbiguousReference},
-		{name: "bare dotted registry port", image: "registry.example.com:5000", want: tunnelImageAmbiguousReference},
-		{name: "dotted slashless name with tag", image: "gcr.io:v1", want: tunnelImageAmbiguousReference},
-		{name: "numeric tag without registry host", image: "qurl-connector:5000", want: tunnelImagePinned},
-		{name: "multi-colon non-latest tag with digest", image: testConnectorImageRepo + ":v1:v2@" + validDigest, want: tunnelImageMalformedReference},
-		{name: "empty tag with digest", image: testConnectorImageRepo + ":@" + validDigest, want: tunnelImageMalformedReference},
-		{name: "malformed digest", image: testConnectorImageRepo + "@notadigest", want: tunnelImageMalformedDigest},
-		{name: "tagged malformed digest", image: testConnectorVersionImage + "@sha256:abc123", want: tunnelImageMalformedDigest},
-		{name: "short sha256 digest", image: testConnectorImageRepo + "@sha256:abc123", want: tunnelImageMalformedDigest},
-		{name: "uppercase sha256 digest", image: testConnectorImageRepo + "@sha256:" + strings.Repeat("A", 64), want: tunnelImageUppercaseDigest},
-		{name: "nameless digest", image: "@" + validDigest, want: tunnelImageMalformedDigest},
-		{name: "digest with sha256 image name", image: "sha256@" + validDigest, want: tunnelImageMalformedDigest},
-		{name: "bare registry digest", image: "localhost:5000@" + validDigest, want: tunnelImageMalformedDigest},
-		{name: "non-numeric registry port with digest", image: "host:abc/qurl-connector@" + validDigest, want: tunnelImageMalformedDigest},
-		{name: "invalid latest-like registry port with digest", image: "registry.example.com:latest/qurl-connector@" + validDigest, want: tunnelImageMalformedDigest},
-		{name: "trailing slash digest", image: "localhost:5000/@" + validDigest, want: tunnelImageMalformedDigest},
-		{name: "leading slash digest", image: "/@" + validDigest, want: tunnelImageMalformedDigest},
-		{name: "bare dotted registry digest", image: "registry.example.com@" + validDigest, want: tunnelImageMalformedDigest},
-		{name: "bare sha256 digest", image: validDigest, want: tunnelImageMalformedDigest},
-		{name: "uppercase bare sha256 digest", image: "SHA256:" + strings.Repeat("a", 64), want: tunnelImageMalformedDigest},
-		{name: "short bare sha256 digest", image: "sha256:abc123", want: tunnelImageMalformedDigest},
-		{name: "at sign in non-digest suffix", image: testConnectorImageRepo + ":weird@tag", want: tunnelImageMalformedDigest},
-	}
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			if got := classifyTunnelImagePin(tc.image); got != tc.want {
-				t.Fatalf("classifyTunnelImagePin(%q) = %v, want %v", tc.image, got, tc.want)
-			}
-		})
-	}
-}
-
-func FuzzClassifyTunnelImagePinRejectsLatest(f *testing.F) {
-	for _, seed := range []string{
-		"",
-		"qurl-connector",
-		testConnectorImageRepo,
-		"registry.example.com:5000/layervai/qurl-connector",
-		"sha256",
-	} {
-		f.Add(seed)
-	}
-	validDigest := "sha256:" + strings.Repeat("a", 64)
-	f.Fuzz(func(t *testing.T, fragment string) {
-		cases := []string{
-			fragment + ":latest",
-			fragment + ":LATEST",
-			fragment + ":latest@" + validDigest,
-			"ghcr.io/foo:latest/" + fragment + ":v1",
-			"ghcr.io/foo:LATEST/" + fragment + "@" + validDigest,
-		}
-		for _, image := range cases {
-			if got := classifyTunnelImagePin(image); got == tunnelImagePinned {
-				t.Fatalf("classifyTunnelImagePin(%q) = tunnelImagePinned, want latest-tagged refs rejected", image)
-			}
-		}
-	})
-}
-
 // applyEnv writes every oauthEnvKeys entry — empty when absent from kvs
 // — so the test doesn't depend on what was inherited from the shell.
 // t.Setenv handles per-test cleanup.
